requirements.txt
# setting up the virtual enviroment inside our initial project file.

python 3.10 
pip install virtualenv
python -m venv tmenv
tmenv\Scripts\activate
pip install django==2.2 
# create a folder with our django project
# inside django project folder
django-admin startproject taskmate
# inside taskmate project folder
python manage.py runserver
python manage.py makemigrations
python manage.py migrate
# create a new app inside taskmate
python manage.py startapp todolist
# create super User inside taskmate folder
python manage.py createsuperuser